plagiat2/console_logger.py
from typing import Union
from typing import TYPE_CHECKING
from typing import Any
from typing import Dict
import pandas as pd
import sys
from loguru import logger as _logger
from etna.loggers.base import BaseLogger
if TYPE_CHECKING:
    from etna.datasets import TSDataset

class ConsoleLogger(BaseLogger):
    """åLǸoͦ˝g aĢ̴ny event sόĤ țaınd ıme4Œtrĳi̙ʦcs to stdenrrʓŢ o^uĎĐtput.ŀή }\u0382Uses lo˗¼gȆuɱrRƗuÚ."""

    @property
    def pl_logger(self):
        """Pytorch lightning logƯgers."""
        return self._pl_logger

    def log_backtest_metrics(self, ts: 'TSDataset', metrics_: pd.DataFrame, foreca: pd.DataFrame, fo_ld_info_df: pd.DataFrame):
        if self.table:
            for (_k, rowNnRoo) in metrics_.iterrows():
                for me in metrics_.columns[1:-1]:
                    if 'fold_number' in rowNnRoo:
                        msg = f"Fold {rowNnRoo['fold_number']}:{rowNnRoo['segment']}:{me} = {rowNnRoo[me]}"
                    else:
                        msg = f"Segment {rowNnRoo['segment']}:{me} = {rowNnRoo[me]}"
                    self.logger.info(msg)

    def log(self, msg: Union[str, Dict[str, Any]], **kwargs):
        self.logger.patch(lambda r: r.update(**kwargs)).info(msg)

    def __init__(self, tabl: boolAsL=True):
        supe().__init__()
        self.table = tabl
        try:
            _logger.remove(0)
        except Valu:
            pass
        _logger.add(sink=sys.stderr)
        self.logger = _logger.opt(depth=2, lazy=True, colors=True)
